refactor: add datagrid models folder
